@@ -24,76 +24,69 @@
     "url": "https://github.com/squidfunk/mkdocs-material.git"
   },
   "scripts": {
-    "build": "npm run clean && npx webpack --mode production",
-    "clean": "npx rimraf material",
-    "lint": "npm run lint:ts && npm run lint:scss",
-    "lint:scss": "npx stylelint \"src/assets/**/*.scss\"",
-    "lint:ts": "npx tslint -p tsconfig.json \"src/**/*.ts\"",
-    "start": "npx webpack --mode development --watch"
+    "build": "run-s build:*",
+    "build:clean": "rimraf material",
+    "build:tools": "ts-node -T tools/build --optimize",
+    "clean": "rimraf material",
+    "lint": "run-p lint:*",
+    "lint:scss": "stylelint \"src/assets/**/*.scss\"",
+    "lint:ts": "eslint --cache \"src/**/*.ts\"",
+    "start": "ts-node -T tools/build --watch"
   },
   "dependencies": {
     "clipboard": "^2.0.6",
     "escape-html": "^1.0.3",
     "focus-visible": "^5.2.0",
+    "fuzzaldrin-plus": "^0.6.0",
     "lunr": "^2.3.9",
     "lunr-languages": "^1.4.0",
-    "rxjs": "^7.0.0-beta.9"
+    "rxjs": "^7.0.0-beta.10"
   },
   "devDependencies": {
-    "@fortawesome/fontawesome-free": "^5.15.1",
-    "@mdi/svg": "^5.8.55",
-    "@primer/octicons": "^11.2.0",
+    "@fortawesome/fontawesome-free": "^5.15.2",
+    "@mdi/svg": "^5.9.55",
+    "@primer/octicons": "^12.0.0",
     "@types/clipboard": "^2.0.1",
-    "@types/copy-webpack-plugin": "^6.4.0",
     "@types/escape-html": "1.0.0",
-    "@types/event-hooks-webpack-plugin": "^2.2.0",
+    "@types/fuzzaldrin-plus": "^0.6.1",
     "@types/html-minifier": "^4.0.0",
     "@types/lunr": "^2.3.3",
-    "@types/mini-css-extract-plugin": "^1.2.2",
-    "@types/node": "^14.14.20",
-    "@types/ramda": "^0.27.34",
+    "@types/node": "^14.14.31",
     "@types/resize-observer-browser": "^0.1.5",
-    "@types/webpack": "^4.41.25",
-    "@types/webpack-assets-manifest": "^3.0.1",
-    "autoprefixer": "10.2.1",
-    "copy-webpack-plugin": "^6.4.1",
-    "css-loader": "^5.0.1",
-    "event-hooks-webpack-plugin": "^2.2.0",
-    "expose-loader": "^1.0.3",
+    "@types/sass": "^1.16.0",
+    "@typescript-eslint/eslint-plugin": "^4.15.1",
+    "@typescript-eslint/parser": "^4.15.1",
+    "autoprefixer": "^10.2.4",
+    "chokidar": "^3.5.1",
+    "cssnano": "^4.1.10",
+    "esbuild": "^0.8.50",
+    "eslint": "^7.20.0",
+    "eslint-plugin-eslint-comments": "^3.2.0",
+    "eslint-plugin-import": "^2.22.1",
+    "eslint-plugin-jsdoc": "^32.1.0",
+    "eslint-plugin-no-null": "^1.0.2",
     "github-types": "^1.0.0",
     "gitlab": "^14.2.2",
     "html-minifier": "^4.0.0",
-    "ignore-emit-webpack-plugin": "2.0.3",
-    "imagemin-webpack-plugin": "^2.4.2",
     "material-design-color": "^2.3.2",
     "material-shadows": "^3.0.1",
-    "mini-css-extract-plugin": "^1.3.3",
-    "postcss": "^8.2.4",
+    "npm-run-all": "^4.1.5",
+    "postcss": "^8.2.5",
     "postcss-inline-svg": "^5.0.0",
-    "postcss-loader": "^4.1.0",
-    "postcss-svgo": "^4.0.2",
-    "preact": "^10.5.9",
-    "ramda": "^0.27.1",
+    "preact": "^10.5.12",
     "rimraf": "^3.0.2",
-    "sass": "^1.32.2",
-    "sass-loader": "^10.1.0",
-    "stylelint": "^13.8.0",
+    "sass": "^1.32.8",
+    "stylelint": "^13.11.0",
     "stylelint-config-rational-order": "^0.1.2",
+    "stylelint-config-recommended": "^3.0.0",
     "stylelint-config-standard": "^20.0.0",
-    "stylelint-order": "^4.1.0",
-    "stylelint-scss": "^3.18.0",
-    "ts-loader": "^8.0.14",
+    "stylelint-scss": "^3.19.0",
+    "svgo": "^2.0.3",
+    "tiny-glob": "^0.2.8",
     "ts-node": "^9.1.1",
-    "tsconfig-paths-webpack-plugin": "^3.3.0",
-    "tslib": "^2.1.0",
-    "tslint": "^6.1.3",
-    "tslint-sonarts": "^1.9.0",
-    "typescript": "^4.1.3",
-    "webpack": "^4.44.2",
-    "webpack-assets-manifest": "3.1.1",
-    "webpack-cli": "^4.3.1"
+    "typescript": "^4.1.5"
   },
   "engines": {
-    "node": ">= 12"
+    "node": ">= 14"
   }
 }
