@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016-2020 Martin Donath <martin.donath@squidfunk.com>
+ * Copyright (c) 2016-2021 Martin Donath <martin.donath@squidfunk.com>
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to
@@ -20,7 +20,7 @@
  * IN THE SOFTWARE.
  */
 
-import "lunr"
+import lunr from "lunr"
 
 import { Search, SearchIndexConfig } from "../../_"
 import {
@@ -72,7 +72,7 @@ let index: Search
  *
  * @param config - Search index configuration
  *
- * @return Promise resolving with no result
+ * @returns Promise resolving with no result
  */
 async function setupSearchLanguages(
   config: SearchIndexConfig
@@ -91,7 +91,7 @@ async function setupSearchLanguages(
   /* Add scripts for languages */
   const scripts = []
   for (const lang of config.lang) {
-    if (lang === "ja") scripts.push(`${base}/tinyseg.min.js`)
+    if (lang === "ja") scripts.push(`${base}/tinyseg.js`)
     if (lang !== "en") scripts.push(`${base}/min/lunr.${lang}.min.js`)
   }
 
@@ -116,7 +116,7 @@ async function setupSearchLanguages(
  *
  * @param message - Source message
  *
- * @return Target message
+ * @returns Target message
  */
 export async function handler(
   message: SearchMessage
@@ -148,6 +148,10 @@ export async function handler(
  * Worker
  * ------------------------------------------------------------------------- */
 
+/* @ts-ignore - expose Lunr.js in global scope, or stemmers will not work */
+self.lunr = lunr
+
+/* Handle messages */
 addEventListener("message", async ev => {
   postMessage(await handler(ev.data))
 })
