@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016-2020 Martin Donath <martin.donath@squidfunk.com>
+ * Copyright (c) 2016-2021 Martin Donath <martin.donath@squidfunk.com>
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to
@@ -20,12 +20,13 @@
  * IN THE SOFTWARE.
  */
 
+import { translation } from "~/_"
 import {
   SearchDocument,
   SearchMetadata,
   SearchResult
-} from "integrations/search"
-import { h, translate, truncate } from "utilities"
+} from "~/integrations/search"
+import { h, truncate } from "~/utilities"
 
 /* ----------------------------------------------------------------------------
  * Helper types
@@ -46,14 +47,14 @@ const enum Flag {
 /**
  * Render a search document
  *
- * @param section - Search document
+ * @param document - Search document
  * @param flag - Render flags
  *
- * @return Element
+ * @returns Element
  */
 function renderSearchDocument(
   document: SearchDocument & SearchMetadata, flag: Flag
-) {
+): HTMLElement {
   const parent = flag & Flag.PARENT
   const teaser = flag & Flag.TEASER
 
@@ -84,7 +85,7 @@ function renderSearchDocument(
         }
         {teaser > 0 && missing.length > 0 &&
           <p class="md-search-result__terms">
-            {translate("search.result.term.missing")}: {...missing}
+            {translation("search.result.term.missing")}: {...missing}
           </p>
         }
       </article>
@@ -100,13 +101,13 @@ function renderSearchDocument(
  * Render a search result
  *
  * @param result - Search result
- * @param threshold - Score threshold
  *
- * @return Element
+ * @returns Element
  */
 export function renderSearchResult(
-  result: SearchResult, threshold: number = Infinity
-) {
+  result: SearchResult
+): HTMLElement {
+  const threshold = result[0].score
   const docs = [...result]
 
   /* Find and extract parent article */
@@ -130,8 +131,8 @@ export function renderSearchResult(
       <details class="md-search-result__more">
         <summary tabIndex={-1}>
           {more.length > 0 && more.length === 1
-            ? translate("search.result.more.one")
-            : translate("search.result.more.other", more.length)
+            ? translation("search.result.more.one")
+            : translation("search.result.more.other", more.length)
           }
         </summary>
         {...more.map(section => renderSearchDocument(section, Flag.TEASER))}
