@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016-2020 Martin Donath <martin.donath@squidfunk.com>
+ * Copyright (c) 2016-2021 Martin Donath <martin.donath@squidfunk.com>
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to
@@ -20,7 +20,7 @@
  * IN THE SOFTWARE.
  */
 
-import { h } from "utilities"
+import { h } from "~/utilities"
 
 /* ----------------------------------------------------------------------------
  * Functions
@@ -31,11 +31,9 @@ import { h } from "utilities"
  *
  * @param table - Table element
  *
- * @return Element
+ * @returns Element
  */
-export function renderTable(
-  table: HTMLTableElement
-) {
+export function renderTable(table: HTMLElement): HTMLElement {
   return (
     <div class="md-typeset__scrollwrap">
       <div class="md-typeset__table">
